+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### v0.13.0.pre
+
+* features
+  * Add payment method definitions methods
+    * `PaymentMethodDefinitions#all`
+    * `PaymentMethodDefinitions#create`
+    * `PaymentMethodDefinitions#delete`
+    * `PaymentMethodDefinitions#find`
+    * `PaymentMethodDefinitions#update`
+  * Add possibility to log request `headers` and `bodies`
+
 ### v0.12.1.pre
 
 * bug-fixes
@@ -14,7 +25,7 @@
     * `Locations#update`
 
 * enhancements
-  * Use `autoload` instead of `require` 
+  * Use `autoload` instead of `require`
 
 ### v0.11.1.pre